Rename dynamic_mixer.rs
